refactor: 이름 검증로직 위치 변경
